renamed app.js to index.js, relocated app sass file to styles folder.
