+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.scss';
-import App from './App/App';
+import App from './App/index';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
