+++ src/styles/_variables.scss
@@ -1,1 +1,1 @@
-$favColor: red;
+$favColor: beige;
